SALEOR-3992 Move new tests out of the deprecated test file
