+++ saleor/graphql/payment/tests/deprecated/test_checkout_payment_create.py
@@ -1,7 +1,4 @@
-from uuid import uuid4
-
 import graphene
-import pytest
 
 from .....checkout import calculations
 from .....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
@@ -130,91 +127,3 @@ def test_checkout_add_payment_both_token_and_id_given(
     assert len(data["errors"]) == 1
     assert not data["payment"]
     assert data["errors"][0]["code"] == PaymentErrorCode.GRAPHQL_ERROR.name
-
-
-def test_create_partial_payments(
-    user_api_client, checkout_without_shipping_required, address
-):
-    # given
-    checkout = checkout_without_shipping_required
-    checkout.billing_address = address
-    checkout.save()
-
-    manager = get_plugins_manager()
-    lines = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(checkout, lines, [], manager)
-    total = calculations.checkout_total(
-        manager=manager, checkout_info=checkout_info, lines=lines, address=address
-    )
-
-    def _variables():
-        return {
-            "token": checkout.token,
-            "input": {
-                "gateway": DUMMY_GATEWAY,
-                "token": uuid4().hex,
-                "amount": total.gross.amount,
-                "partial": True,
-            },
-        }
-
-    # when
-    response_1 = user_api_client.post_graphql(CREATE_PAYMENT_MUTATION, _variables())
-    response_2 = user_api_client.post_graphql(CREATE_PAYMENT_MUTATION, _variables())
-    content_1 = get_graphql_content(response_1)
-    content_2 = get_graphql_content(response_2)
-    data_1 = content_1["data"]["checkoutPaymentCreate"]
-    data_2 = content_2["data"]["checkoutPaymentCreate"]
-
-    # then
-    assert data_1["payment"]
-    assert data_2["payment"]
-    assert checkout.payments.filter(is_active=True, partial=True).count() == 2
-
-
-@pytest.mark.parametrize("first_payment_is_partial", [True, False])
-def test_create_subsequent_full_payment(
-    user_api_client,
-    checkout_without_shipping_required,
-    address,
-    first_payment_is_partial,
-):
-    # given
-    checkout = checkout_without_shipping_required
-    checkout.billing_address = address
-    checkout.save()
-
-    manager = get_plugins_manager()
-    lines = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(checkout, lines, [], manager)
-    total = calculations.checkout_total(
-        manager=manager, checkout_info=checkout_info, lines=lines, address=address
-    )
-
-    def _variables(partial):
-        return {
-            "token": checkout.token,
-            "input": {
-                "gateway": DUMMY_GATEWAY,
-                "token": uuid4().hex,
-                "amount": total.gross.amount,
-                "partial": partial,
-            },
-        }
-
-    # when
-    response_1 = user_api_client.post_graphql(
-        CREATE_PAYMENT_MUTATION, _variables(first_payment_is_partial)
-    )
-    response_2 = user_api_client.post_graphql(
-        CREATE_PAYMENT_MUTATION, _variables(False)
-    )
-    content_1 = get_graphql_content(response_1)
-    content_2 = get_graphql_content(response_2)
-    data_1 = content_1["data"]["checkoutPaymentCreate"]
-    data_2 = content_2["data"]["checkoutPaymentCreate"]
-
-    # then
-    assert data_1["payment"]
-    assert data_2["payment"]
-    assert checkout.payments.filter(is_active=True).count() == 1
